refactor: order table package 분리
