+++ src/main/java/kitchenpos/ordertable/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.ordertable.domain;
 
 import static kitchenpos.common.exception.ExceptionMessage.*;
 
@@ -12,6 +12,7 @@ import javax.persistence.Id;
 
 import kitchenpos.common.domain.Empty;
 import kitchenpos.common.exception.BadRequestException;
+import kitchenpos.order.domain.NumberOfGuests;
 
 @Entity
 public class OrderTable {
