+++ src/main/java/kitchenpos/ordertable/domain/TableUngroupEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.ordertable.domain;
 
 public class TableUngroupEvent {
