+++ src/main/java/kitchenpos/ordertable/repository/OrderTableRepository.java
@@ -1,10 +1,10 @@
-package kitchenpos.order.repository;
+package kitchenpos.ordertable.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
     List<OrderTable> findAllByIdIn(List<Long> orderTableIds);
