+++ src/main/java/kitchenpos/ordertable/repository/TableGroupRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.order.repository;
+package kitchenpos.ordertable.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.order.domain.TableGroup;
+import kitchenpos.ordertable.domain.TableGroup;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
 }
